+++ src/app/(pages)/login/page.tsx
@@ -1,7 +1,7 @@
 "use client";
 import { InputWrapper } from "@/components/InputWrapper";
 import { Card, Input, Button, Typography } from "@/lib/material-tailwind";
-import { useAuth } from "@/app/hooks/useAuth";
+import { useLogin } from "@/app/hooks/useLogin";
 import { useMutateSnackbar } from "@/app/hooks/useMutateSnackbar";
 import { snackbarState } from "@/app/store/snackbar";
 import { auth, provider } from "@/firebase/client";
@@ -21,7 +21,7 @@ type FormInputsType = {
 
 export default function Page() {
   const router = useRouter();
-  const { isLoading, login } = useAuth();
+  const { isLoading, login } = useLogin();
   const [isShowPassword, setIsShowPassword] = useState(false);
   const setActiveSnackbar = useSetAtom(snackbarState);
