+++ src/app/hooks/useLogin.tsx
@@ -1,11 +1,9 @@
 import { zMessageResponse } from "@/types";
 import { useSetAtom } from "jotai";
-import { useRouter } from "next/navigation";
 import { useCallback, useState } from "react";
 import { snackbarState } from "../store/snackbar";
 
-export const useAuth = () => {
-  const router = useRouter();
+export const useLogin = () => {
   const [isLoading, setIsLoading] = useState<boolean>(false);
   const setActiveSnackbar = useSetAtom(snackbarState);
 
@@ -46,39 +44,5 @@ export const useAuth = () => {
     [setActiveSnackbar]
   );
 
-  const logout = useCallback(async () => {
-    setIsLoading(true);
-    try {
-      const response = await fetch(
-        `${process.env.NEXT_PUBLIC_BASE_URL}/api/auth/logout`,
-        {
-          method: "POST",
-          headers: {
-            "Content-Type": "application/json",
-          },
-        }
-      );
-      if (!response.ok) {
-        throw new Error(response.statusText);
-      }
-      const { message } = zMessageResponse.parse(await response.json());
-      console.log(message);
-    } catch (error) {
-      setActiveSnackbar({
-        isOpen: true,
-        message: "ログアウトに失敗しました",
-        loading: false,
-        isError: true,
-      });
-      if (error instanceof Error) {
-        console.error(error.message);
-      } else {
-        console.error(error);
-      }
-    }
-
-    setIsLoading(false);
-  }, [setActiveSnackbar]);
-
-  return { isLoading, login, logout };
+  return { isLoading, login };
 };
